fix line display when zoomout in maps and refacto zoomToFeature function
